Rename main.py modules to table.py
