+++ sqlite_store/jsonstore/__init__.py
@@ -1,6 +1,6 @@
 """Public API for the :mod:`sqlite_store.jsonstore` package."""
 
-from .main import (
+from .table import (
     create_json_table,
     insert_json,
     insert_json_auto_hash,
