+++ sqlite_store/jsonstore/table.py
