+++ tests/test_jsonstore.py
@@ -6,7 +6,7 @@ import hashlib
 import random
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from sqlite_store.jsonstore.main import (
+from sqlite_store.jsonstore.table import (
     create_json_table,
     insert_json,
     insert_json_auto_hash,
