+++ tests/test_jsonstore_fts.py
@@ -3,7 +3,7 @@ import sys
 import sqlite3
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from sqlite_store.jsonstore.main import create_json_table, insert_json
+from sqlite_store.jsonstore.table import create_json_table, insert_json
 from sqlite_store.jsonstore.fts import create_json_fts
