+++ tests/test_objectstore.py
@@ -5,7 +5,7 @@ import json
 import hashlib
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from sqlite_store.objectstore.main import (
+from sqlite_store.objectstore.table import (
     create_object_table,
     insert_object,
     insert_object_auto_hash,
