+++ tests/test_objectstore_view.py
@@ -3,7 +3,7 @@ import sys
 import sqlite3
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from sqlite_store.objectstore.main import (
+from sqlite_store.objectstore.table import (
     create_object_table,
     insert_object,
 )
